JsonExtension class also moved to engine
