+++ engine/core/src/main/java/es/eucm/ead/engine/utils/JsonExtension.java
@@ -34,7 +34,7 @@
  *      You should have received a copy of the GNU Lesser General Public License
  *      along with eAdventure.  If not, see <http://www.gnu.org/licenses/>.
  */
-package es.eucm.ead.schemax;
+package es.eucm.ead.engine.utils;
 
 /**
  * Convenient utility for handling Json extensions in {@code String}s and
